requirements.txt
pytest-playwright
pandas
python-docx
os
urllib3
pillow